requirements.txt
streamlit
python-dotenv
langchain-google-genai
langchain-openai
youtube-transcript-api
requests
pydantic